put json data into the list

+++ src/ie/gmit/os/FitnessRecord.java
@@ -1,14 +1,14 @@
 package ie.gmit.os;
 
 public class FitnessRecord {
-	private int index;
+	private long index;
 	private String mode;
-	private float duration;
+	private double duration;
 	
-	public int getIndex() {
+	public long getIndex() {
 		return index;
 	}
-	public void setIndex(int index) {
+	public void setIndex(long index) {
 		this.index = index;
 	}
 	public String getMode() {
@@ -17,10 +17,10 @@ public class FitnessRecord {
 	public void setMode(String mode) {
 		this.mode = mode;
 	}
-	public float getDuration() {
+	public double getDuration() {
 		return duration;
 	}
-	public void setDuration(float duration) {
+	public void setDuration(double duration) {
 		this.duration = duration;
 	}
 }
